More refactoring to rationalize view vs. edit mode.
