+++ src/hi/apps/entity/edit/entity_type_transition_handler.py
@@ -12,10 +12,11 @@ from django.http import HttpRequest, HttpResponse
 import hi.apps.common.antinode as antinode
 from hi.apps.location.location_manager import LocationManager
 from hi.apps.location.models import LocationView
-from .entity_manager import EntityManager
+from hi.apps.entity.entity_manager import EntityManager
+from hi.apps.entity.models import Entity
+from hi.apps.entity.forms import EntityForm, EntityAttributeRegularFormSet
+
 from .enums import EntityTransitionType
-from .models import Entity
-from .forms import EntityForm, EntityAttributeRegularFormSet
 
 logger = logging.getLogger(__name__)
