+++ src/hi/apps/entity/urls.py
@@ -9,7 +9,6 @@ urlpatterns = [
              views.EntityEditView.as_view(), 
              name='entity_edit'),
 
-
     re_path( r'^status/(?P<entity_id>\d+)$', 
              views.EntityStatusView.as_view(), 
              name='entity_status' ),
@@ -18,21 +17,17 @@ urlpatterns = [
              views.EntityStateHistoryView.as_view(), 
              name='entity_state_history' ),
 
-
     re_path( r'^attribute/upload/(?P<entity_id>\d+)$', 
              views.EntityAttributeUploadView.as_view(), 
-             name='entity_attribute_upload'),
-    
+             name='entity_attribute_upload' ),
 
-    re_path( r'^entity/(?P<entity_id>\d+)/attribute/(?P<attribute_id>\d+)/history/inline/$', 
+    re_path( r'^attribute/history/(?P<entity_id>\d+)/(?P<attribute_id>\d+)/$', 
              views.EntityAttributeHistoryInlineView.as_view(), 
-             name='entity_attribute_history_inline'),
+             name='entity_attribute_history_inline' ),
     
-    re_path( 
-        r'^entity/(?P<entity_id>\d+)/attribute/(?P<attribute_id>\d+)/restore/inline/(?P<history_id>\d+)/$', 
-        views.EntityAttributeRestoreInlineView.as_view(), 
-        name='entity_attribute_restore_inline'
-    ),
+    re_path( r'^attribute/restore/(?P<entity_id>\d+)/(?P<attribute_id>\d+)/(?P<history_id>\d+)/$', 
+             views.EntityAttributeRestoreInlineView.as_view(), 
+             name='entity_attribute_restore_inline' ),
 
     re_path( r'^edit/', include('hi.apps.entity.edit.urls' )),
