+++ src/hi/apps/location/edit/forms.py
@@ -4,15 +4,12 @@ import os
 from django import forms
 from django.conf import settings
 
-from hi.apps.attribute.forms import AttributeForm, AttributeUploadForm
-from hi.apps.attribute.enums import AttributeValueType
 from hi.apps.common.svg_forms import SvgDecimalFormField, SvgFileForm
 from hi.apps.common.svg_models import SvgItemPositionBounds
 
 from hi.apps.location.enums import LocationViewType, SvgStyleName
 from hi.apps.location.models import (
     Location,
-    LocationAttribute,
     LocationView,
 )
 
@@ -72,63 +69,6 @@ class LocationEditForm( forms.ModelForm ):
             'order_id': forms.NumberInput( attrs={'class': 'form-control'} ),
             'svg_view_box_str': forms.TextInput( attrs={'class': 'form-control', 'readonly': True } ),
         }
-
-        
-class LocationAttributeForm( AttributeForm ):
-    class Meta( AttributeForm.Meta ):
-        model = LocationAttribute
-        
-        
-LocationAttributeFormSet = forms.inlineformset_factory(
-    Location,
-    LocationAttribute,
-    form = LocationAttributeForm,
-    extra = 1,
-    max_num = 100,
-    absolute_max = 100,
-    can_delete = True,
-)
-
-
-class LocationAttributeUploadForm( AttributeUploadForm ):
-    class Meta( AttributeUploadForm.Meta ):
-        model = LocationAttribute
-
-
-# V2 Modal Forms - only name field for modal editing
-class LocationV2EditForm( forms.ModelForm ):
-    """
-    Location edit form for V2 modal - only includes name field.
-    Geometry fields (svg_view_box_str, order_id) are edited separately.
-    """
-    
-    class Meta:
-        model = Location
-        fields = ('name',)
-        widgets = {
-            'name': forms.TextInput(attrs={'class': 'form-control'})
-        }
-
-
-class RegularAttributeBaseFormSet(forms.BaseInlineFormSet):
-    """Base formset that automatically excludes FILE attributes for regular attribute editing"""
-    
-    def get_queryset(self):
-        """Override to automatically filter out FILE attributes"""
-        queryset = super().get_queryset()
-        return queryset.exclude(value_type_str=str(AttributeValueType.FILE))
-
-
-LocationAttributeRegularFormSet = forms.inlineformset_factory(
-    Location,
-    LocationAttribute,
-    form = LocationAttributeForm,
-    formset = RegularAttributeBaseFormSet,
-    extra = 1,
-    max_num = 100,
-    absolute_max = 100,
-    can_delete = True,
-)
 
 
 class LocationViewAddForm(forms.Form):
