+++ src/hi/apps/location/location_edit_response_renderer.py
@@ -15,7 +15,7 @@ import hi.apps.common.antinode as antinode
 from hi.constants import DIVID
 from .location_edit_form_handler import LocationEditFormHandler
 from .models import Location, LocationAttribute
-from .edit.forms import LocationV2EditForm, LocationAttributeRegularFormSet
+from .forms import LocationAttributeRegularFormSet, LocationModalEditForm
 from .location_attribute_edit_context import LocationAttributeEditContext
 
 
@@ -34,7 +34,7 @@ class LocationEditResponseRenderer:
 
     def build_template_context( self,
                                 location                     : Location,
-                                location_form                : LocationV2EditForm,
+                                location_form                : LocationModalEditForm,
                                 file_attributes              : QuerySet[LocationAttribute],
                                 regular_attributes_formset   : LocationAttributeRegularFormSet,
                                 success_message              : Optional[str] = None,
@@ -44,7 +44,7 @@ class LocationEditResponseRenderer:
         
         Args:
             location: Location instance
-            location_form: LocationV2EditForm instance
+            location_form: LocationModalEditForm instance
             file_attributes: QuerySet of file attributes
             regular_attributes_formset: LocationAttributeRegularFormSet instance
             success_message: Optional success message for display
@@ -80,7 +80,7 @@ class LocationEditResponseRenderer:
     def render_update_fragments( self,
                                  request                     : HttpRequest,
                                  location                    : Location,
-                                 location_form               : Optional[LocationV2EditForm] = None,
+                                 location_form               : Optional[LocationModalEditForm] = None,
                                  regular_attributes_formset  : Optional[LocationAttributeRegularFormSet] = None,
                                  success_message             : Optional[str] = None,
                                  error_message               : Optional[str] = None,
@@ -183,14 +183,14 @@ class LocationEditResponseRenderer:
             self,
             request                     : HttpRequest,
             location                    : Location,
-            location_form               : LocationV2EditForm,
+            location_form               : LocationModalEditForm,
             regular_attributes_formset  : LocationAttributeRegularFormSet ) -> 'antinode.Response':
         """Render error response using antinode helpers - multiple target replacement.
         
         Args:
             request: HTTP request object
             location: Location instance
-            location_form: LocationV2EditForm instance with validation errors
+            location_form: LocationModalEditForm instance with validation errors
             regular_attributes_formset: FormSet instance with validation errors
             
         Returns:
